Creation of featureCard part one
